test(domain): separate validated analysis capabilities
